refactor: move error classes to errors.py
