+++ cryosparc/spec.py
@@ -212,56 +212,6 @@ A result slot specification for the slots=... argument.
 """
 
 
-class SlotsValidation(TypedDict):
-    """
-    Data from validation error when specifying external result input/output slots.
-
-    :meta private:
-    """
-
-    type: Datatype
-    valid: List[SlotSpec]
-    invalid: List[Datafield]
-    valid_dtypes: List[str]
-
-
-class InvalidSlotsError(ValueError):
-    """
-    Raised by functions that accept slots arguments, and CryoSPARC reports that
-    given slots are not valid.
-    """
-
-    def __init__(self, caller: str, validation: SlotsValidation):
-        type = validation["type"]
-        valid_slots = validation["valid"]
-        invalid_slots = validation["invalid"]
-        valid_dtypes = validation["valid_dtypes"]
-        msg = "\n".join(
-            [
-                f"Unknown {type} slot dtype(s): {', '.join(s['dtype'] for s in invalid_slots)}. "
-                "Only the following slot dtypes are valid:",
-                "",
-            ]
-            + [f" - {t}" for t in valid_dtypes]
-            + [
-                "",
-                "If adding a dynamic result such as alignments_class_#, specify a "
-                "slots=... argument with a full data field specification:",
-                "",
-                f"    {caller}(... , slots=[",
-                "        ...",
-            ]
-            + [f"        {repr(s)}," for s in valid_slots]
-            + [
-                "        {'dtype': '<INSERT HERE>', 'prefix': '%s', 'required': True}," % s["dtype"]
-                for s in invalid_slots
-            ]
-            + ["        ...", "    ])"]
-        )
-
-        return super().__init__(msg)
-
-
 class InputSlot(TypedDict):
     """
     Dictionary entry in Job document's ``input_slot_groups.slots`` property.
