+++ docs/api/errors.rst
@@ -1,0 +1,6 @@
+cryosparc.errors
+================
+
+.. automodule:: cryosparc.errors
+    :autosummary:
+    :members:
